backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22

python-multipart>=0.0.5

spacy>=3.5,<4

pandas>=1.5
numpy>=1.23,<2.0
scikit-learn>=1.0

sentence-transformers>=2.2.0
torch>=1.13.1

pdfplumber>=0.7.5
python-docx>=0.8.11

pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=9.0.0
